Refactor isCosmosEndpoint to minimal platform-specific logic
