+++ sdk/tables/data-tables/src/TableTransaction.ts
@@ -30,7 +30,7 @@ import type { TableClientLike } from "./utils/internalModels.js";
 import type { TableServiceErrorOdataError } from "./generated/index.js";
 import { cosmosPatchPolicy } from "./cosmosPathPolicy.js";
 import { getTransactionHeaders } from "#platform/utils/transactionHeaders";
-import { isCosmosEndpoint } from "#platform/utils/isCosmosEndpoint";
+import { isCosmosEndpoint } from "./utils/isCosmosEndpointUtil.js";
 import { tracingClient } from "./utils/tracing.js";
 
 /**
