+++ sdk/tables/data-tables/src/utils/isCosmosEndpoint-react-native.mts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export { isCosmosEndpoint } from "./isCosmosEndpoint-browser.mjs";
+export { getAzuriteAccounts } from "./isCosmosEndpoint-browser.mjs";
