+++ sdk/tables/data-tables/src/utils/isCosmosEndpointUtil.ts
@@ -1,0 +1,34 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { getAzuriteAccounts } from "#platform/utils/isCosmosEndpoint";
+
+export function isCosmosEndpoint(url: string): boolean {
+  const parsedURL = new URL(url);
+  if (parsedURL.hostname.indexOf(".table.cosmosdb.") !== -1) {
+    return true;
+  }
+
+  if (parsedURL.hostname.indexOf(".table.cosmos.") !== -1) {
+    return true;
+  }
+
+  // Azurite emulator IP-style URL for table?
+  if (
+    (parsedURL.hostname === "localhost" || parsedURL.hostname === "127.0.0.1") &&
+    parsedURL.pathname.startsWith("/devstoreaccount1")
+  ) {
+    return false;
+  }
+
+  const azuriteAccounts = getAzuriteAccounts()?.split(":");
+  if (azuriteAccounts?.[0] && parsedURL.hostname.includes(azuriteAccounts[0])) {
+    return false;
+  }
+
+  if (parsedURL.hostname === "localhost") {
+    return true;
+  }
+
+  return false;
+}
